+++ app/gui/view/src/lib.rs
@@ -31,7 +31,6 @@
 
 #[allow(clippy::option_map_unit_fn)]
 pub mod code_editor;
-pub mod component_browser;
 pub mod debug_mode_popup;
 pub mod project;
 pub mod project_list;
@@ -40,6 +39,7 @@ pub mod searcher;
 pub mod status_bar;
 pub mod window_control_buttons;
 
+pub use ide_view_component_browser as component_browser;
 pub use ide_view_documentation as documentation;
 pub use ide_view_graph_editor as graph_editor;
 pub use welcome_screen;
